+++ src/command_modules/azure-cli-configure/azure/cli/command_modules/configure/_consts.py
@@ -39,3 +39,4 @@ MSG_PROMPT_TELEMETRY = '\nMicrosoft would like to collect anonymous Azure CLI us
                        'about how you use Azure CLI.  To update your choice, run "az configure" ' \
                        'again.\nSelect y to enable data collection.'
 MSG_PROMPT_FILE_LOGGING = '\nWould you like to enable logging to file?'
+MSG_PROMPT_LOGGING_COLORS = '\nWould you like to colorize your log messages on screen?'